feat: cardplay analysis in the engine + a WebAssembly build (#1)
* feat: bring cardplay analysis into the engine, and add a WebAssembly build

The cardplay analysis lived in bridge-solver-service/src/analyse_play.rs, which
made it reachable only over HTTP. Moving it into the engine lets a browser run
the same analysis directly, and means the service and any other caller share one
implementation rather than drifting.

analyse_play.rs moves here verbatim behind a new `play-analysis` feature, off by
default so callers wanting only a double-dummy result do not pull serde and
sha2. Its 13 tests came with it and pass unchanged.

wasm/ exposes the three operations the service does, with matching request and
response JSON so a client can swap transport without reshaping data:

  Analyzer::dd_table      <- POST /dd            <- par::solve_dd_table
  Analyzer::dd_play       <- POST /dd/play       <- analyse_play::running_trace
  Analyzer::dd_play_node  <- POST /dd/play/node  <- analyse_play::node_alternatives

The service persists position values in SQLite; there is no database in a
browser, so Analyzer keeps the same prefix-keyed map in memory for its lifetime.
Stepping through a hand then costs one solve per newly reached position.

Also fixes a wasm blocker in solve_with_caches_and_partial: it read
Instant::now() unconditionally, but only used it for the optional [PERF] line.
std::time::Instant has no wasm32 implementation and traps, so every browser
solve would have aborted — verified by building a probe and running it under
node, which panicked with "unreachable". The clock is now read only when the
diagnostic is on, and compiled out entirely for wasm32. That also removes a
clock read from the native hot path.

Verified against Bridge Base Online's own server-computed table for a real
board: all 20 cells of dd_table match, contract_tricks agrees, and the trace
finds 5 costed errors with seat attribution. In node the full 41-card trace
takes ~128ms and a repeat is 0.2ms off the cache.


* fix(search): avoid u8 underflow computing relative pattern-cache bounds

`(result.ns_tricks - ns_tricks_won)` subtracted two u8 values before casting to
i8. When the search returns fewer NS tricks than NS has already won — normal
whenever the remaining tricks all go to EW — that underflows: a debug build
panics with "attempt to subtract with overflow", and a release build wraps to a
large positive u8 that then casts to a nonsense negative i8, poisoning the
pattern cache bounds.

Cast each operand to i8 first so the subtraction is signed and the negative
result is the intended one. `relative_tricks` is already used as a signed value
by both Bounds::new arms.


* chore: restore a green CI

The Lint job has been failing on main since a14e712 (2026-07-15): `cargo fmt
--all -- --check` reports diffs at src/par.rs:291 and :322, where the par test
case table was hand-wrapped rather than left to rustfmt. Two things follow from
that: nobody was told, because a red Lint job on main sends no signal anywhere,
and every later branch inherits the failure and looks broken on arrival.

Two fixes, both in par.rs:

* rustfmt applied to the test case tables. This is only reflow, no case
  changed. It is a big diff for a small reason; keeping it in its own commit
  means the surrounding work stays readable.
* `best.map_or(true, ..)` -> `best.is_none_or(..)`, which clippy now requires.
  Semantically identical, and `is_none_or` has been stable since 1.82 so every
  toolchain the CI matrix uses has it.

After this, `cargo fmt --all -- --check`, `cargo clippy --all-targets --
-D warnings` and the test suite are all clean locally.


* chore(ci): standardize the invocations, and cover the new feature and member

Matches the other repos: `cargo fmt --all -- --check`, `cargo clippy
--workspace --all-targets`, `cargo test --workspace`. No toolchain pinning.

Two repo-specific adjustments beyond the common form:

* `--workspace` now reaches the new wasm/ member, and the test and second clippy
  passes name `play-analysis` so the 13 analyse_play tests are actually run —
  with default features they do not compile and were silently uncovered.
* Not `--all-features`. That switches on the behaviour-altering debug flags
  (no_tricks_pruning, no_fast_tricks, ...) together, which changes solver
  results and fails the suite.

`cli` is deliberately left off the test job: it builds the CLI binaries, whose
tests::test_replaces_existing_dd_tags has been stale since a14e712 added par
calculation — it still asserts OptimumScore/ParContract are absent, which that
commit made them no longer be. CI never caught it because the cli feature's
tests had never run. Left for a deliberate fix rather than changed here, with a
note in the workflow.


* docs: roadmap for the browser work

Captures what is done, what is left, and the things that were measured or
verified so a later session does not re-derive them: the wasm timings, the
verification against BBO's own DD table, and the three platform traps that
compile fine and then fail (Instant, rayon, reqwest::blocking).

Workstreams are marked independent where they are, so the remaining pieces —
LIN input, the Pages site, the BBO extension, folding EDGAR onto this engine —
can be picked up in parallel. LIN input blocks the site and the extension; the
rest do not block each other.

Also records two known issues found along the way: the stale
test_replaces_existing_dd_tags in the CLI binary, and that the Lint job sat red
on main for two weeks without notifying anyone.


---------

+++ src/search.rs
@@ -653,7 +653,7 @@ impl<'a> Search<'a> {
 
         // Pattern cache store (matching C++ common_bounds_cache)
         if !NO_TT.load(Ordering::Relaxed) && !pattern_cutoff {
-            let relative_tricks = (result.ns_tricks - ns_tricks_won) as i8;
+            let relative_tricks = result.ns_tricks as i8 - ns_tricks_won as i8;
             let bounds = if (result.ns_tricks as i8) < beta {
                 Bounds::new(0, relative_tricks)
             } else {
